requirements.txt
application-properties==0.6.0
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
Columnar==1.4.1
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8-absolute-import==1.0.0.1
flake8-commas==2.1.0
flake8-quotes==3.3.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.1.0
pymarkdownlnt==0.9.11
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
